refactor : 역할 클래스 변경으로 테스트 코드 위치 변경
