+++ src/test/java/nextstep/ladder/PersonsTest.java
@@ -5,10 +5,7 @@ import ladder.util.StringUtil;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
-
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -37,25 +34,5 @@ public class PersonsTest {
         assertThatThrownBy(() -> Persons.of(inputPersons))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage(StringUtil.EMPTY_WORDS_EXCEPTION);
-    }
-
-    @ParameterizedTest
-    @DisplayName("사람별 보상값 정상적으로 가져오는지 테스트")
-    @CsvSource(value = {"pobi:보상3","honux:보상1","jk:보상2"}, delimiter = ':')
-    void prizeResultTest(String targetName, String targetPrize) {
-        Persons persons = Persons.of("pobi,honux,jk");
-        LadderConstants ladderConstants = new LadderConstants();
-        LadderGameSetting ladderGameSetting = LadderGameSetting.of(persons, ladderConstants.prizes);
-        List<ResultPrize> resultPrizes = ladderConstants.ladder.findPrize(ladderGameSetting);
-
-        String prizeValue = resultPrizes.stream()
-                                    .filter(resultPrize -> resultPrize.checkEqaulName(targetName))
-                                    .map(resultPrize -> resultPrize.getPrizeValue())
-                                    .findAny()
-                                    .orElse("");
-
-        assertThat(prizeValue).isEqualTo(targetPrize);
-
-
     }
 }
